# openapi/current.json
{
  "components": {
    "schemas": {
...
      "CreateFocusPlanRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "recommendationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendationid"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/FocusPlanSegmentInput"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/FocusPlanSource"
          },
          "template": {
            "default": false,
            "title": "Template",
            "type": "boolean"
          },
          "workItemId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workitemid"
          }
        },
        "required": [
          "source",
          "name",
          "segments"
        ],
        "title": "CreateFocusPlanRequest",
        "type": "object"
      },
      "CreateRecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "availableTimeSeconds": {
            "anyOf": [
              {
                "maximum": 43200.0,
                "minimum": 300.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availabletimeseconds"
          },
          "preferredStrategy": {
            "$ref": "#/components/schemas/RecommendationStrategy",
            "default": "auto"
          },
          "workItemId": {
            "format": "uuid",
            "title": "Workitemid",
            "type": "string"
          }
        },
        "required": [
          "workItemId"
        ],
        "title": "CreateRecommendationRequest",
        "type": "object"
      },
...
      "DurationMode": {
        "enum": [
          "fixed",
          "flexible",
          "open"
        ],
        "title": "DurationMode",
        "type": "string"
      },
...
      "FocusPlan": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recommendationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendationid"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/FocusPlanSegment"
            },
            "title": "Segments",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/FocusPlanSource"
          },
          "template": {
            "title": "Template",
            "type": "boolean"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workItemId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workitemid"
          }
        },
        "required": [
          "id",
          "workItemId",
          "recommendationId",
          "source",
          "name",
          "template",
          "segments",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "title": "FocusPlan",
        "type": "object"
      },
      "FocusPlanPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FocusPlan"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "FocusPlanPage",
        "type": "object"
      },
      "FocusPlanRecommendation": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "engineVersion": {
            "title": "Engineversion",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/FocusPlanSnapshot"
          },
          "provenance": {
            "$ref": "#/components/schemas/RecommendationProvenance"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/RecommendationReason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "workItemId": {
            "format": "uuid",
            "title": "Workitemid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workItemId",
          "engineVersion",
          "provenance",
          "plan",
          "reasons",
          "createdAt"
        ],
        "title": "FocusPlanRecommendation",
        "type": "object"
      },
      "FocusPlanSegment": {
        "additionalProperties": false,
        "properties": {
          "durationMode": {
            "$ref": "#/components/schemas/DurationMode"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "index": {
            "minimum": 0.0,
            "title": "Index",
            "type": "integer"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "kind": {
            "$ref": "#/components/schemas/FocusSegmentKind"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "linkedWorkItemId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedworkitemid"
          },
          "maxSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxseconds"
          },
          "minSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minseconds"
          },
          "optional": {
            "default": false,
            "title": "Optional",
            "type": "boolean"
          },
          "targetSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetseconds"
          }
        },
        "required": [
          "kind",
          "durationMode",
          "id",
          "index"
        ],
        "title": "FocusPlanSegment",
        "type": "object"
      },
      "FocusPlanSegmentInput": {
        "additionalProperties": false,
        "properties": {
          "durationMode": {
            "$ref": "#/components/schemas/DurationMode"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "kind": {
            "$ref": "#/components/schemas/FocusSegmentKind"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "linkedWorkItemId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedworkitemid"
          },
          "maxSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxseconds"
          },
          "minSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minseconds"
          },
          "optional": {
            "default": false,
            "title": "Optional",
            "type": "boolean"
          },
          "targetSeconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetseconds"
          }
        },
        "required": [
          "kind",
          "durationMode"
        ],
        "title": "FocusPlanSegmentInput",
        "type": "object"
      },
      "FocusPlanSnapshot": {
        "additionalProperties": false,
        "properties": {
          "segments": {
            "items": {
              "$ref": "#/components/schemas/FocusPlanSegmentInput"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "required": [
          "segments"
        ],
        "title": "FocusPlanSnapshot",
        "type": "object"
      },
      "FocusPlanSource": {
        "enum": [
          "manual",
          "recommendation",
          "template"
        ],
        "title": "FocusPlanSource",
        "type": "string"
      },
      "FocusSegmentKind": {
        "enum": [
          "work",
          "break",
          "long_break",
          "retrieval",
          "review",
          "planning",
          "transition"
        ],
        "title": "FocusSegmentKind",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "hasMore",
          "nextCursor"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "RecommendationProvenance": {
        "enum": [
          "bootstrap",
          "preference_informed",
          "history_informed"
        ],
        "title": "RecommendationProvenance",
        "type": "string"
      },
      "RecommendationReason": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "label": {
            "$ref": "#/components/schemas/RecommendationReasonLabel"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label",
          "message"
        ],
        "title": "RecommendationReason",
        "type": "object"
      },
      "RecommendationReasonLabel": {
        "enum": [
          "evidence_backed_general",
          "evidence_mixed",
          "product_heuristic",
          "personal_pattern",
          "user_preference"
        ],
        "title": "RecommendationReasonLabel",
        "type": "string"
      },
      "RecommendationStrategy": {
        "enum": [
          "auto",
          "continuous",
          "structured",
          "flexible"
        ],
        "title": "RecommendationStrategy",
        "type": "string"
      },
...
      "UpdateFocusPlanRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FocusPlanSegmentInput"
                },
                "maxItems": 100,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "template": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template"
          }
        },
        "title": "UpdateFocusPlanRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Client-independent REST API for Taskiller.",
    "title": "Taskiller API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/focus-plan-recommendations": {
      "post": {
        "operationId": "createFocusPlanRecommendation",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlanRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Focus Plan Recommendation",
        "tags": [
          "Focus"
        ]
      }
    },
    "/api/v1/focus-plan-recommendations/{recommendationId}": {
      "get": {
        "operationId": "getFocusPlanRecommendation",
        "parameters": [
          {
            "in": "path",
            "name": "recommendationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recommendationid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlanRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Focus Plan Recommendation",
        "tags": [
          "Focus"
        ]
      }
    },
    "/api/v1/focus-plans": {
      "get": {
        "operationId": "listFocusPlans",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "workItemId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workitemid"
            }
          },
          {
            "in": "query",
            "name": "templateOnly",
            "required": false,
            "schema": {
              "default": false,
              "title": "Templateonly",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlanPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Focus Plans",
        "tags": [
          "Focus"
        ]
      },
      "post": {
        "operationId": "createFocusPlan",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFocusPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Focus Plan",
        "tags": [
          "Focus"
        ]
      }
    },
    "/api/v1/focus-plans/{focusPlanId}": {
      "delete": {
        "operationId": "deleteFocusPlan",
        "parameters": [
          {
            "in": "path",
            "name": "focusPlanId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Focusplanid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete Focus Plan",
        "tags": [
          "Focus"
        ]
      },
      "get": {
        "operationId": "getFocusPlan",
        "parameters": [
          {
            "in": "path",
            "name": "focusPlanId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Focusplanid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Focus Plan",
        "tags": [
          "Focus"
        ]
      },
      "patch": {
        "operationId": "updateFocusPlan",
        "parameters": [
          {
            "in": "path",
            "name": "focusPlanId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Focusplanid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFocusPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Focus Plan",
        "tags": [
          "Focus"
        ]
      }
    },
...
  }
}